refactor(codegen): extrai analysis/captures.rs de func.rs
Move 14 funcoes (637 linhas) de captura/rename/promote para
crates/rts-codegen/src/codegen/lower/analysis/captures.rs:
- collect_local_decls / collect_decls_in_stmt
- collect_captures_in_body / scan_*_for_arrows / collect_captured_from_arrow
- collect_idents_used_in_{stmt,expr}
- rename_ident_in_{stmt,expr} / rename_uses_in_body
- make_sync_param_to_global / promote_local_to_global
- extract_class_owner

func.rs: 9161 -> 8524 linhas. Suite verde (todos os 11 crates).
