+++ crates/rts-codegen/src/codegen/lower/analysis/mod.rs
@@ -3,4 +3,5 @@
 //! Submodulos folha (sem deps cruzadas entre si) extraidos de `func.rs`.
 
 pub mod address_taken;
+pub mod captures;
 pub mod types;
